Renames SelectHiveAcc to HiveSelectAcc
