+++ src/components/HiveLogin.vue
@@ -1,0 +1,11 @@
+<template>
+  <div>
+    <HiveSelectAcc />
+  </div>
+</template>
+
+<script setup>
+import HiveSelectAcc from "components/HiveSelectAcc.vue"
+</script>
+
+<style lang="scss" scoped></style>
